+++ frontend/php/edit.php
@@ -26,13 +26,15 @@ include "header.php";
 						ID:<input class=\"input_id\" type=\"text\" name=\"id_edit\" value=\"$id\">
 						Description:<input type=\"text\" name=\"description_edit\" value=\"$description\"></div>";
 				echo "<div id=\"checkbox_script\">";
-				for ($x = 2; $x < $count_scripts; $x++) {
-					if(is_link("./machines/$mac/".$scripts[$x])) {
-						echo $scripts[$x].":<input type='checkbox' name='".$scripts[$x]."' value='1' checked>";
-					} elseif ($scripts[$x] == 'custom') {
-						echo "<button class='custom'>Custom</button>";
-					} else {
-						echo $scripts[$x].":<input type='checkbox' name='".$scripts[$x]."' value='1'>";
+				for ($x = 0; $x < $count_scripts; $x++) {
+					if ($scripts[$x] != "." && $scripts[$x] != "..") {
+						if(is_link("./machines/$mac/".$scripts[$x])) {
+							echo $scripts[$x].":<input type='checkbox' name='".$scripts[$x]."' value='1' checked>";
+						} elseif ($scripts[$x] == 'custom') {
+							echo "<button class='custom'>Custom</button>";
+						} else {
+							echo $scripts[$x].":<input type='checkbox' name='".$scripts[$x]."' value='1'>";
+						}
 					}
 				}
 				echo "</div>";
@@ -47,18 +49,20 @@ include "header.php";
 			echo "<br />";
 			echo "$file/$dir[2]";
 			echo "<br />";
-			for ($i = 2; $i < $count; $i++) {
-				if(!unlink("$file/".$dir[$i])) {
-					echo "Couldn't delete file.";
-				} else {
-					unlink($file/$dir[$i]);
+			for ($i = 0; $i < $count; $i++) {
+				if ($dir[$i] != "." && $dir[$i] != "..") {
+					if(!unlink("$file/".$dir[$i])) {
+						echo "Couldn't delete file.";
+					} else {
+						unlink($file/$dir[$i]);
+					}
 				}
-			}
-			if(!rmdir($file)) {
-				echo "Couldn't delete dir";
-			} else {
-				rmdir($file);
-				header('Location: ./manage.php');
+				if(!rmdir($file)) {
+					echo "Couldn't delete dir";
+				} else {
+					rmdir($file);
+					header('Location: ./manage.php');
+				}
 			}
 		}
 		?>
