+++ frontend/php/edit_script.php
@@ -13,9 +13,11 @@ if ($submit == "Submit") {
 	if (is_link("./scripts/$file")) {
 		unlink("./scripts/$file");
 	}
-	for ($x = 2; $x < $count_machines; $x++) {
-		if (is_link("./machines/$machine[$x]/$file")) {
-			unlink("./machines/$machine[$x]/$file");
+	for ($x = 0; $x < $count_machines; $x++) {
+		if ($machine[$x] != "." && $machine[$x] != "..") {
+			if (is_link("./machines/$machine[$x]/$file")) {
+				unlink("./machines/$machine[$x]/$file");
+			}
 		}
 	}
 }
