+++ frontend/php/save_changes.php
@@ -8,15 +8,17 @@ $count_scripts = count($scripts);
 
 $data = "$mac|$id|$description";
 
-for ($x = 2; $x < $count_scripts; $x++) {
-	if ($scripts[$x] != 'custom') {
-		$post = $_POST[$scripts[$x]];
-	}
-	if ($post == "1") {
-		// Step back up two dirs for the symlink.
-		symlink("../../scripts/$scripts[$x]", "./machines/$mac/$scripts[$x]");
-	} elseif ($scripts[$x] != 'custom') {
-		unlink("./machines/$mac/$scripts[$x]");
+for ($x = 0; $x < $count_scripts; $x++) {
+	if ($scripts[$x] != "." && $scripts[$x] != "..") {
+		if ($scripts[$x] != 'custom') {
+			$post = $_POST[$scripts[$x]];
+		}
+		if ($post == "1") {
+			// Step back up two dirs for the symlink.
+			symlink("../../scripts/$scripts[$x]", "./machines/$mac/$scripts[$x]");
+		} elseif ($scripts[$x] != 'custom') {
+			unlink("./machines/$mac/$scripts[$x]");
+		}
 	}
 }
